move methods from __init__.py to auxiliar.py
